requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
livekit-agents==0.8.4
livekit-plugins-openai==0.6.4
